Break out an abstract channel factory
